fix: separate auth layout from main layout by using route group
Move main pages into (main) route group so auth routes don't inherit
Header/Footer from root layout. Root layout now only provides html,
body, and ThemeProvider wrapper.

Fixes #157
