+++ app/(main)/explore/[eventId]/page.tsx
