+++ app/(main)/explore/page.tsx
@@ -1,6 +1,6 @@
-// import Header from "../components/Header";
-// import Footer from "../components/Footer";
-import MainContent from "../components/explore/MainContent";
+// import Header from "../../components/Header";
+// import Footer from "../../components/Footer";
+import MainContent from "../../components/explore/MainContent";
 import { getAllPublicEvents } from "@/lib/dataFetching";
 import { Metadata } from "next";
