+++ app/(main)/layout.tsx
@@ -1,0 +1,16 @@
+import Header from "@/app/components/Header";
+import Footer from "@/app/components/Footer";
+
+export default function MainLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <Header />
+      {children}
+      <Footer />
+    </>
+  );
+}
