+++ app/(main)/news/[slug]/page.tsx
